# docs/service/openapi.yml
openapi: 3.0.1
info:
  title: Project Jano
  description: 'This is the API for the microservices of the Project Jano'
  termsOfService: https://github.com/project-jano
  contact:
    email: ezequiel.aceto+project-jano@gmail.com
  license:
    name: Mozilla Public License
    url: https://www.mozilla.org/en-US/MPL/
  version: 2.0.4
servers:
  - url: http://localhost:8080/
...
paths:
...
  /v2/users/{userId}/messages/decode:
    post:
      tags:
        - Users
      summary: Turns a SecuredMessage into a Text Plain message
      description: Decrypts a SecuredPayload and verify its integrity using the given parameters.
      operationId: DecodeSecureMessage
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        - in: query
          name: deviceId
          required: false
          schema:
            type: string
          example: "abcdef00-0011-3214-abcd-174426614abc"
          description: | 
            The Device ID where the message was secured. If not specified, the default device is used.
            
            **WARNING**: Using the default device in a multi-device (per user) environment requires that you keep which device is the default up to date.          
        - in: query
          name: keyId
          required: false
          schema:
            type: string
          example: "default"
          description: |
            The Key ID of the private key used to secure the message. If not specified, the default key (default) is used. 
            
            **WARNING**: Only set a value for this field if you are completely sure of what you are doing. One credential, the default one, per device and user is enough for the 99.9% of the applications.          
      requestBody:
        description: Secured payload and related information in order to decrypt it and verify its integrity.
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/DecodeSecuredMessageRequest'
        required: true
      responses:
        201:
          description: successfully signed the certificate
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  anyOf:
                    - $ref: '#/components/schemas/Payload'

        400:
          description: Invalid request
        401:
          description: Invalid API Key
        404:
          description: UserId and/or deviceId not found
  /v2/users/{userId}/push-notifications/secure:
    post:
      tags:
        - Users
      summary: Secures a push notification
      description: Encrypts and Signs a message of type push notification (title, subtitle and payload) with credentials available for the given parameters. In any case, it will return an array with at least 1 element.
      operationId: SecurePushNotificationForUser
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        - in: query
          name: deviceIds
          required: false
          description: | 
            (Comma-separated value) list of devices to use. If not specified, the default device is used if exists. If keyword **all** is used, all the available devices will be used.
            
            **WARNING**: Using the default device in a multi-device (per user) environment requires that you keep which device is the default up to date.
          schema:
            type: string
          example: "abcdef00-0011-3214-abcd-174426614abc, fegcab99-2200-1234-dcba-174426614abc" 
        - in: query
          name: completeOutput
          required: false
          description: If *true* returns the complete output of the encrypted text and its signature.
          schema:
            type: boolean
            default: true
          allowEmptyValue: true
        - in: query
          name: splittedOutput
          required: false
          description: If *true* returns the splitted output of the encrypted text and its signature.
          schema:
            type: boolean
            default: false
          allowEmptyValue: true
        - in: query
          name: keyId
          required: false
          description: |
            The Key ID of the private key used to secure the message. If not specified, the default key ("default") is used. 
            
            **WARNING**: Only set a value for this field if you are completely sure of what you are doing. One credential, the default one, per device and user is enough for the 99.9% of the applications.
          schema:
            type: string
          example: "default"
          allowEmptyValue: false
          
      requestBody:
        description: Push notification
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/SecurePushNotificationRequest'
        required: true
      responses:
        201:
          description: Successfully created a SecuredMessage based on the message
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  anyOf:
                    - $ref: '#/components/schemas/SecuredPushNotification'

        400:
          description: Invalid request
        401:
          description: Invalid API Key
        404:
          description: UserId and/or devices not found
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    DecodeSecuredMessageRequest:
      type: object
      required:
        - keyId
        - securedMessage
        - signature
      properties:
        keyId:
          type: string
          description: Id of the (private) key in the user's device
          example: "default"
        securedMessage:
          type: string
          description: Secured message
          example: "JlCDYGryyXkLXBByWobzhWjm2fu3AuOZ0sEe+I4+LRQcx+8ZSJplMfonJ5aRH0BjxazosXGrUNNMnrvPkGUrXlbMqaom0DYkaVDWFP3/adg9NeVUxR69X978qHvugzFUXJfUqPuQryz1rE2G3jGUo91DKrK2XVVASDdkOUABufVOm0qfE+WWyJzdodUMFKxp16FIQ580T0bTaUeCRMY5C6qAG7TX547qO2KcWCnhSftRCfid13Ohc3qmTzGfQaFxtd93mlACvrlZRpI4o7veqYOAEhR+/w/rwS5yH6v903W5kCFt9jTyKSfesAPMgaJor4gC4N5GRkG2V3nv+zx1+ZKF/ZFN511U9I84bXu59HoWKBXW+7sCbCD2hMkL1wqWVOXO5CxvtCcxxeNBsUN0sr8/v4ct+8Wzn/+qc7222DEN4eEBs69k6Rx4aDTaRdH6Ig1ZvUFa3YP34T78iywGSd0K7BpdQjTrHmr6x80p/+HmSvY255OyDfkiaynU4gEH6BsUMANxOEVsSZiOA84zdX1ggQ6cqhaM7UFylXDHlwS6Iz/YVrOlILI2V8DJD8uZ8L5NuP7fE9vA01HOjwB1tiOYbH5MPEEln+QWkwSdNzuyRKK3JjttIoxAgHqDm31UGxVkupIaq4M9/9kPSDo5Z4tE5TACq91m4KAiIqQW+s0="
        signature:
          type: string
          description: Signature of the payload field using the hash function
          example: "cefb2414c887cafec94bf8a7d6f8ec0ac98cefd59085959618b771d087c3f0351023d377cf88b490300d416fa1f8efd5bba1f2f8a2c70efa142f196713d48282238d16a6c74e1c6d66e903ca913a4219827fded0856d002ac8c97f8b19a4ffe27245fd783ad295234a6b05891421a235ee99310fac22e5eee9e1d630e3a0de7e45362515f750aba1d71553ce33b77ef991042a9a24c1ff692059a608de5f60b4ff4ef57e8cf85f9a0dfea6af21724809513d221b1361247e26b55840919231b9ecbb0f212814a28a6bdf569b2ec3fbc573f122f4b881e95dba4eb7622be86f83a1a8610fef4d0d98ba736a06dad568c9fe04409c290f73063ce6091d8169ba30cd9f0bc00d4a4f69ff05e421ca9863feedef29a5d4b1c8fbfff59da92567f64f422d258f6dcc4c3b9376cd1431ac8f301f8b5074fc5ec62786d3c35b709198bd5f51aed94bc569cd24baa4b84d50284ee18a911a6213b3a96d35229003cb2236acc3eab39a02b506f06417659437ddd259c7976a37482196a0874f208379cf2ef1beb568976d3ab6bec1552f35e2ab3f3b53948de7bf6fc61d908559a87cc0492f4580cd70440609d899e91439693caa52d9d198ad431999fdbd22f9ddb328563df5a9d8dd58df00e4a337baaf0ba97ad5abbe3562d59dafbb3cbbe153401256147b56f772ff688d1e745a63441416c6f238a89c901787ef11d47f376f27b751"
...
    SecurePushNotificationRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the Push Notification
          example: "Hello, world!"
        subtitle:
          type: string
          description: Subtitle of the Push Notification
          example: "This is the subtitle"          
        payload:
          type: string
          description: A JSON (string) of parameters
          example: "{\"key-1\":\"value-a\", \"key-2\":\"value-b\"}"
    Payload:
      type: object
      required:
        - timestamp
        - fingerprint
        - message
      properties:
        timestamp:
          type: integer
          description: Unix timestamp of the server when securing the payload in mS.
          example: 1628687400000
        fingerprint:
          type: string
          example: "8176747f6127efc931cc9f7e8d5c216ae34de6f7e1528508f2c8ec3abd411e88"
          description: The fingerprint of the service instance that encrypted this message
        message:
          type: string
          description: plain text message
          example: "{\"key\":\"value\"}"
    SecuredPushNotification:
      type: object
      required:
        - keyId
        - deviceId
        - payload
        - signature
      properties:
        keyId:
          type: string
          description: Id of the (private) key in the user's device
          example: "default"
        deviceId:
          type: string
          description: Id of the user's device
          example: "test-device-1"
        payload:
          type: object
          required: 
            - complete
            - splitted
          properties:
            complete:
              type: string
              description: Secured push notification in complete string
              example: "TtBSkwmcyIs1CRpw4y34YCXi6BFkzcprhRV5MBZ6y/3ctc+6zidzsgGQbkK+ZgEzXHVE4sFy07W/3iBVcKLMAbAjoyxLwfd6YoCBtN+QlQsjWQuK5Ji4K2Rj1w/oaAiAuM8usnfTE39CM23Sa6jtbFBMQ5wOtKWjMg26weGsfNISbnNSI5nSwmNhbDB4v/DiwAr2bnQVh5pkJiISa6Ew14HlUpefn7Ctm7pNb9H7s71Efq9Md/idOk1Vdcz3pKiYxx3E9ungyeanMYJVa5o5zrx6DA3Sqzt5I+IcK3X3lv0BACQybSCdaMFLuAWyhcg2r8r7IEL99ApB2z+o2RhRAbWVo7G8hblPdoyYeKmvDd675sI1ng4iVVqAtZthejXYIPeVQ2s1SIQmGOCVWnD3Kifae9xgToXQTZ6S8aBydq35NTfs6BbBvrUWOe2zfDc45AKDYCVHNj9iyhAzPW4TI4Nr5oMsGOHTvdJm9h0xBJoSz2Qlcr0snluijjVwTo36G09k8inUjnkqzd/H/bLDlFzXphJYJf6VdwHwcb4fT8/DDL/fmNyKWDeOkuUWNQwOq0XXBq5XzUcw0acvmslNKSks28lRvro46YEUyHW2UXJYzP/nz0i4lDWUwdA2cgHoHC5XWWC6RFJReIBuim8RrqjcI9BMaRBqjCCpHna82b0="
            splitted:
              type: array
              items:
                type: string
              description: Payload splitted in strings of 200 chars max (Firebase Limitation)
              example:
                [
                  "TtBSkwmcyIs1CRpw4y34YCXi6BFkzcprhRV5MBZ6y/3ctc+6zidzsgGQbkK+ZgEzXHVE4sFy07W/3iBVcKLMAbAjoyxLwfd6YoCBtN+QlQsjWQuK5Ji4K2Rj1w/oaAiAuM8usnfTE39CM23Sa6jtbFBMQ5wOtKWjMg26weGsfNISbnNSI5nSwmNhbDB4v/DiwAr2bnQV",
                  "h5pkJiISa6Ew14HlUpefn7Ctm7pNb9H7s71Efq9Md/idOk1Vdcz3pKiYxx3E9ungyeanMYJVa5o5zrx6DA3Sqzt5I+IcK3X3lv0BACQybSCdaMFLuAWyhcg2r8r7IEL99ApB2z+o2RhRAbWVo7G8hblPdoyYeKmvDd675sI1ng4iVVqAtZthejXYIPeVQ2s1SIQmGOCV",
                  "WnD3Kifae9xgToXQTZ6S8aBydq35NTfs6BbBvrUWOe2zfDc45AKDYCVHNj9iyhAzPW4TI4Nr5oMsGOHTvdJm9h0xBJoSz2Qlcr0snluijjVwTo36G09k8inUjnkqzd/H/bLDlFzXphJYJf6VdwHwcb4fT8/DDL/fmNyKWDeOkuUWNQwOq0XXBq5XzUcw0acvmslNKSks",
                  "28lRvro46YEUyHW2UXJYzP/nz0i4lDWUwdA2cgHoHC5XWWC6RFJReIBuim8RrqjcI9BMaRBqjCCpHna82b0="
                ]
        signature:
          type: object
          description: Signature of the encrypted payload field using the hash function
          required: 
            - complete
            - splitted
          properties:
            complete:
              type: string
              description: Complete signature (in one line string)
              example: "IrRYqzklcE/yv/ZD/EbXApQxSHvEaezedLB0gagD2iMnpdO8VUPteF/L53RQGbNFHkAOzKMVUs6Jrhvs0I0WAQVzWP19tv6DCnJMDp2dlk1DJASOGS8JiUrRK4PHNsbwoVXh3Nh3aquaCpcy0n96fQ2ADk8avrDCWirTq2A9PlKL6DlXO/qYXsJGB14RxHPCOOut/Vw0svQIm+LWwoLeL+qWi85Rd0cDFe4FT1fo5c3U/wgrkO7Qh/wS81I/dQpueI9d4c2V0N/v2c7nyjQxF1+25BGdgLjMWoY4p83D3KMxVhsNv1ZbnIF6vkJ7haENex/E94e2/kGlhjnBIyiFeDOYV8Qc9a3uOgSTXbDd4duqAgBRmrftdr1vWdJUD0312IfxOLHZ9ju/cXM15ndUgSNI+e4dfE3mYE87vApG7gPZC/nDm5wD6UUOOxMMUhbP162pEB+yfZ5q3Iry6BcA/Ou6ygyMSajp5nslr2xywWOtL65fSBlqnM0K0/fJatg+mEiJs2kJhtrPf3MU2AaXgEkTqURBb5KJa+9Wq2cGN4J4vjcHLqC1lVzwx5/iSks27vdno8kbxYAmGn/YxAQ4ic0pzK6zFgqjepwHyVxhah5FbT0CP/iOqQKk7PrFEzFzeIH487CeYXP2DSNy9USMiTcamE1urc8+yCs+8lOQRD4="
            splitted:
              type: array
              items:
                type: string
              description: Signature splitted in strings of 200 chars max (Firebase Limitation)
              example:
                [
                  "IrRYqzklcE/yv/ZD/EbXApQxSHvEaezedLB0gagD2iMnpdO8VUPteF/L53RQGbNFHkAOzKMVUs6Jrhvs0I0WAQVzWP19tv6DCnJMDp2dlk1DJASOGS8JiUrRK4PHNsbwoVXh3Nh3aquaCpcy0n96fQ2ADk8avrDCWirTq2A9PlKL6DlXO/qYXsJGB14RxHPCOOut/Vw0",
                  "svQIm+LWwoLeL+qWi85Rd0cDFe4FT1fo5c3U/wgrkO7Qh/wS81I/dQpueI9d4c2V0N/v2c7nyjQxF1+25BGdgLjMWoY4p83D3KMxVhsNv1ZbnIF6vkJ7haENex/E94e2/kGlhjnBIyiFeDOYV8Qc9a3uOgSTXbDd4duqAgBRmrftdr1vWdJUD0312IfxOLHZ9ju/cXM1",
                  "5ndUgSNI+e4dfE3mYE87vApG7gPZC/nDm5wD6UUOOxMMUhbP162pEB+yfZ5q3Iry6BcA/Ou6ygyMSajp5nslr2xywWOtL65fSBlqnM0K0/fJatg+mEiJs2kJhtrPf3MU2AaXgEkTqURBb5KJa+9Wq2cGN4J4vjcHLqC1lVzwx5/iSks27vdno8kbxYAmGn/YxAQ4ic0p",
                  "zK6zFgqjepwHyVxhah5FbT0CP/iOqQKk7PrFEzFzeIH487CeYXP2DSNy9USMiTcamE1urc8+yCs+8lOQRD4="
                ]
...
security:
  - basicAuth: [ ]